Clean, Fixes, Add signature and Fabric
Remove unused files;
Clean the style and and strings;
Fixed a crash with the light settings
Add signature for the app (build.gradle not "committing" by now);
Add Fabric with Crashlytics and Answers;
Changed the splash screen (will be changed again in the future).

First version (0.1) send by Fabric for Beta Tests.